+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Fix bug causing incorrect validation in lost item fee policies. Fixes UICIRC-667, UICIRC-668.
 * Add RTL/Jest testing for `ScedulesList` component. Refs UICIRC-601.
 * Add RTL/Jest testing for `OverdueAboutSection` component. Refs UICIRC-595. 
+* Add RTL/Jest testing for `normalize` function in `FinePolicy/utils`. Refs UICIRC-599.
 
 ## [5.1.0] (https://github.com/folio-org/ui-circulation/tree/v5.1.0) (2021-06-14)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v5.0.1...v5.1.0)